fix(editor): Move cURL command to http node params to frontend
