+++ packages/editor-ui/src/stores/ui.store.ts
@@ -544,19 +544,6 @@ export const useUIStore = defineStore(STORES.UI, () => {
 		sidebarMenuCollapsed.value = !sidebarMenuCollapsed.value;
 	};
 
-	const getCurlToJson = async (curlCommand: string) => {
-		const parameters = await curlParserApi.getCurlToJson(rootStore.restApiContext, curlCommand);
-
-		// Normalize placeholder values
-		if (parameters['parameters.url']) {
-			parameters['parameters.url'] = parameters['parameters.url']
-				.replaceAll('%7B', '{')
-				.replaceAll('%7D', '}');
-		}
-
-		return parameters;
-	};
-
 	const goToUpgrade = async (
 		source: CloudUpdateLinkSourceType,
 		utm_campaign: UTMCampaign,
@@ -694,7 +681,6 @@ export const useUIStore = defineStore(STORES.UI, () => {
 		resetSelectedNodes,
 		setCurlCommand,
 		toggleSidebarMenuCollapse,
-		getCurlToJson,
 		goToUpgrade,
 		removeBannerFromStack,
 		dismissBanner,
